imageAi/requirements (1).txt
streamlit
ultralytics
opencv-python
numpy
Pillow
imageai
torch 
torchvision